Se modularizo OrderDetail dividiendose en diferentes components para facilitar mantenimiento y legibilidad
